+++ phoenix-bootloader/src/main.rs
@@ -7,32 +7,25 @@
 
 mod arch;
 
+
 #[macro_use]
 extern crate alloc;
 
 #[macro_use]
 extern crate log;
 
-use core::panic::PanicInfo;
 use core::fmt::Write;
 use alloc::string::String;
-use alloc::vec::Vec;
 
-use uefi::proto::console::text::Output;
-use uefi::{prelude::*, table::boot::*};
-use uefi_services::system_table;
-use uefi::proto::console::gop::GraphicsOutput;
-use uefi::proto::console::gop::{BltOp, BltPixel, BltRegion, FrameBuffer, PixelFormat};
-use uefi::table::boot::BootServices;
-use embedded_graphics::pixelcolor::{Rgb888, RgbColor};
+use uefi::{prelude::*};
 
-const ASCII_INTRO: &str = r"     
+const ASCII_IMAGE: &str = r"                                                                                                          
               ,,\\\                     ///,,       
             (\\\\//                     \\////)      
            (-(__//                       \\__)-)     
          ((-(__||                         ||__)-))    
         ((-(-(_||           ```\__        ||_)-)-))   
-       ((-(-(/(/\\        ''; 9.- `      //\)\)-)-))   
+       ((-(-(/(/\\        ''; 9.--'      //\)\)-)-))   
        (-(-(/(/(/\\      '';;;;-\~      //\)\)\)-)-)   
        ((-(-(/(/(/\======,:;:;:;:,======/\)\)\)-)-))   
        '(((-(/(/(/(//////:%%%%%%%:\\\\\\)\)\)\)-)))`  
@@ -40,44 +33,44 @@ const ASCII_INTRO: &str = r"
          '((-(/(/(/('|||:wwwwwwwww:|||')\)\)\)-))`    
            '((((/(/('uuu:WWWWWWWWW:uuu`)\)\))))`     
              '':::UUUUUU:wwwwwwwww:UUUUUU:::``     
-                  '''''''\uuuuuuuu/``````         
-                         `JJJJJJJJJ`           
-                          LLLLLLLLLL         
-                         ///|||||||\\\       
-                        (/(/(/(^)\)\)\)   
+                   '''''''uuuuuuu``````   
+                         /JJJJJJJ\    
+                        /LLLLLLLLL\   
+                       ///|||||||\\\   
+                      (/(/(/(^)\)\)\)
 ";
+
+const ASCII_INTRO: &str = r"
+ __        __   ___             __   __  
+|__) |__| /  \ |__  |\ | | \_/ /  \ /__` 
+|    |  | \__/ |___ | \| | / \ \__/ .__/ ";
+
 const SPLASH_INTRO: &'static [u8] = include_bytes!("../../other/splash.bmp");
-
-static mut BOOT_SYSTEM_TABLE: Option<SystemTable<Boot>> = None;
-static mut IMAGE: Option<Handle> = None;
-
 
 #[no_mangle]
 extern "C" fn efi_main(
   image: Handle, 
   mut system_table: SystemTable<Boot>, 
-  rev: uefi::table::Revision
 )-> Status{
     uefi_services::init(&mut system_table).expect_success("Failed to initialize utilities");
    
-    system_table.stdout().clear().expect("Failed to clean screen");
-
+    system_table.stdout().clear().expect("Failed to clean screen").unwrap();
+ 
     arch::main(image, &mut system_table);
- 
-    system_table.stdout().clear().expect("Failed to clean screen");
- 
 
-
-    writeln!(system_table.stdout(), "{}", ASCII_INTRO);
-    writeln!(system_table.stdout(), "Phoenix Bootloader v{}, created by {}",env!("CARGO_PKG_VERSION"), env!("CARGO_PKG_AUTHORS"));
+    Status::SUCCESS.is_success();    
     
     loop{}
 }
 
-fn check_revision(rev: uefi::table::Revision){
+pub fn check_revision(rev: uefi::table::Revision, system_table: &mut SystemTable<Boot>){
   let (major, minor) = (rev.major(), rev.minor());
+  
+  let mut buf = String::new();
+    system_table.firmware_vendor().as_str_in_buf(&mut buf).unwrap();
 
-  log::info!("UEFI {}.{}", major, minor / 10);
+  writeln!(system_table.stdout(),"INFO: UEFI {}.{}", major, minor / 10).unwrap();
+  writeln!(system_table.stdout(),"INFO: Firmware Vendor: {}", buf.as_str()).unwrap();
 
   assert!(major >= 2, "Running on an old, unsupported version of UEFI");
   assert!(
@@ -86,4 +79,3 @@ fn check_revision(rev: uefi::table::Revision){
   );
 
 }
-