+++ phoenix-builder/src/pack/aarch64_pack.rs
@@ -144,7 +144,7 @@ pub fn debug_package_files() -> Result<(), Box<dyn Error>> {
     let efi_file = Path::new("phoenix-bootloader/target/aarch64-unknown-uefi/debug/phoenix-bootloader.efi");
 
     let kernel_file = Path::new("phoenix-kernel/target/aarch64-phoenix/debug/phoenix-kernel");
-    let out_path = Path::new("build").join("aarch64");
+    let out_path = Path::new("build").join("aarch64").join("debug");
 
     let fat_path = out_path.join("phoenix.fat");
     let img_path = out_path.join("phoenix.img");
@@ -163,7 +163,7 @@ pub fn release_package_files() -> Result<(), Box<dyn Error>> {
     let efi_file = Path::new("phoenix-bootloader/target/aarch64-unknown-uefi/release/phoenix-bootloader.efi");
 
     let kernel_file = Path::new("phoenix-kernel/target/aarch64-phoenix/release/phoenix-kernel");
-    let out_path = Path::new("build").join("aarch64");
+    let out_path = Path::new("build").join("aarch64").join("release");
 
     let fat_path = out_path.join("phoenix.fat");
     let img_path = out_path.join("phoenix.img");