+++ phoenix-builder/src/pack/x86_pack.rs
@@ -143,7 +143,7 @@ pub fn debug_package_files() -> Result<(), Box<dyn Error>> {
     let efi_file = Path::new("phoenix-bootloader/target/x86_64-unknown-uefi/debug/phoenix-bootloader.efi");
 
     let kernel_file = Path::new("phoenix-kernel/target/x86_64-phoenix/debug/phoenix-kernel");
-    let out_path = Path::new("build").join("x86_64");
+    let out_path = Path::new("build").join("x86_64").join("debug");
 
     let fat_path = out_path.join("phoenix.fat");
     let img_path = out_path.join("phoenix.img");
@@ -162,7 +162,7 @@ pub fn release_package_files() -> Result<(), Box<dyn Error>> {
     let efi_file = Path::new("phoenix-bootloader/target/x86_64-unknown-uefi/release/phoenix-bootloader.efi");
 
     let kernel_file = Path::new("phoenix-kernel/target/x86_64-phoenix/release/phoenix-kernel");
-    let out_path = Path::new("build").join("x86_64");
+    let out_path = Path::new("build").join("x86_64").join("release");
 
     let fat_path = out_path.join("phoenix.fat");
     let img_path = out_path.join("phoenix.img");